moved getModels logic to modelController
